:recycle: [repositories] Move `provide` tests to `test_registry`
